+++ src/Oro/Bundle/CronBundle/Command/CronCommandActivationInterface.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Oro\Bundle\CronBundle\Command;
+
+/**
+ * Represents a CRON command with conditional activation.
+ */
+interface CronCommandActivationInterface
+{
+    /**
+     * Checks whether the CRON command is active (i.e. properly configured, etc).
+     *
+     * @return bool
+     */
+    public function isActive();
+}