+++ src/Oro/Bundle/CronBundle/Tests/Unit/Stub/CronCommandStub.php
@@ -2,10 +2,10 @@
 
 namespace Oro\Bundle\CronBundle\Tests\Unit\Stub;
 
-use Oro\Bundle\CronBundle\Command\CronCommandInterface;
+use Oro\Bundle\CronBundle\Command\CronCommandScheduleDefinitionInterface;
 use Symfony\Component\Console\Command\Command;
 
-class CronCommandStub extends Command implements CronCommandInterface
+class CronCommandStub extends Command implements CronCommandScheduleDefinitionInterface
 {
     /**
      * {@inheritDoc}
@@ -13,13 +13,5 @@ class CronCommandStub extends Command implements CronCommandInterface
     public function getDefaultDefinition(): string
     {
         return '*/1 * * * *';
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function isActive(): bool
-    {
-        return true;
     }
 }